gdb/testsuite: Allow for failure to read some memory addresses
In the gdb.base/examine-backward.exp test script, we check to see if
address zero is readable, and then read memory first forward from
address zero, and then backward from address zero.

The problem is, being able to read address zero does not guarantee that
you'll be able to read from the other end of the address space, and the
test probably shouldn't assume that is the case.

This patch extends the success conditions so that, even if GDB fails to
read memory, so long as the error message indicates that GDB was trying
to access the correct location, then we consider this a pass.  The test
is, I think, trying to show that GDB can compute the correct address
when going backward from zero, being able to access the memory at that
address is secondary.

One further change is that, when we examined the memory at address zero,
the regexp used to match the address expected that the "zero address"
would have two '0' digits as the least significant digits.  As GDB
strips leading zeros from addresses, I assume that this test was
originally written for a target that inserts address flags into the most
significant part of the address, and so, the address zero, would appear
like: '0x<FLAGS>00000000' or similar.  On targets without address flags,
reading address zero just prints the address as '0x0'.  I've tweaked the
address pattern so either case should now pass.

gdb/testsuite/ChangeLog:

	* gdb.base/examine-backward.exp: Be more forgiving of failures to
	read some addresses.